refactor: drop redundant Float64 conversion in PSD scalar call
the vector method accepts AbstractVector{<:Real} and _psd_values already
narrows each query with Float64(xq[i]), so the outer conversion was a no-op

+++ AstroSGWB/src/detector/psd.jl
@@ -69,6 +69,6 @@ function (p::PowerSpectralDensity)(frequencies::AbstractVector{<:Real})
 end
 
 function (p::PowerSpectralDensity)(f::Real)
-    v = p([Float64(f)])
+    v = p([f])
     return v[1]
 end
